+++ integrations/anthropic/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [integrations/anthropic-v5.1.1] - 2026-01-15
+
+### 🐛 Bug Fixes
+
+- None value handling of flattened generation kwargs for AnthropicChatGenerator (#2753)
+
+
 ## [integrations/anthropic-v5.1.0] - 2026-01-13
 
 ### 🚀 Features